feat: polish detail
